shares mock
