+++ services/director-v2/tests/unit/test_modules_osparc_variables.py
@@ -10,15 +10,12 @@ import json
 from collections.abc import AsyncIterable
 from contextlib import asynccontextmanager
 from copy import deepcopy
-from datetime import timedelta
 from unittest.mock import AsyncMock, Mock
 
 import pytest
 from asgi_lifespan import LifespanManager
 from faker import Faker
 from fastapi import FastAPI
-from models_library.api_schemas_webserver.auth import ApiKeyGet
-from models_library.products import ProductName
 from models_library.service_settings_labels import ComposeSpecLabelDict
 from models_library.services import ServiceKey, ServiceVersion
 from models_library.users import UserID
@@ -144,35 +141,6 @@ def mock_user_repo(mocker: MockerFixture, mock_repo_db_engine: None) -> None:
 
 
 @pytest.fixture
-def mock_api_key_manager(mocker: MockerFixture) -> None:
-
-    fake_data = ApiKeyGet.parse_obj(ApiKeyGet.Config.schema_extra["examples"][0])
-
-    async def _create(
-        app: FastAPI,
-        *,
-        product_name: ProductName,
-        user_id: UserID,
-        name: str,
-        expiration: timedelta,
-    ):
-        assert app
-        assert product_name
-        assert user_id
-        assert expiration is None
-
-        fake_data.display_name = name
-        return fake_data
-
-    # mocks RPC interface
-    mocker.patch(
-        "simcore_service_director_v2.modules.osparc_variables._api_auth.get_or_create_api_key_and_secret",
-        side_effect=_create,
-        autospec=True,
-    )
-
-
-@pytest.fixture
 async def fake_app(faker: Faker) -> AsyncIterable[FastAPI]:
     app = FastAPI()
     app.state.engine = AsyncMock()
@@ -188,7 +156,10 @@ async def fake_app(faker: Faker) -> AsyncIterable[FastAPI]:
 
 
 async def test_resolve_and_substitute_session_variables_in_specs(
-    mock_user_repo: None, mock_api_key_manager: None, fake_app: FastAPI, faker: Faker
+    mock_user_repo: None,
+    mock_osparc_variables_api_auth_rpc: None,
+    fake_app: FastAPI,
+    faker: Faker,
 ):
     specs = {
         "product_name": "${OSPARC_VARIABLE_PRODUCT_NAME}",
